4-side padding, refactoring

+++ android/src/main/java/com/reactlibrary/scene/nodes/UiButtonNode.kt
@@ -61,7 +61,7 @@ class UiButtonNode(props: ReadableMap, context: Context) : UiNode(props, context
 
     private fun setTextSize(props: Bundle) {
         if (props.containsKey(PROP_TEXT_SIZE)) {
-            val textSize = props.getDouble(PROP_TEXT_SIZE)
+            val textSize = props.getDouble(PROP_TEXT_SIZE).toFloat()
             val size = Utils.metersToPx(textSize, view.context).toFloat()
             (view as Button).setTextSize(TypedValue.COMPLEX_UNIT_PX, size)
         }
