+++ android/src/main/java/com/reactlibrary/scene/nodes/UiTextNode.kt
@@ -75,7 +75,7 @@ class UiTextNode(props: ReadableMap, context: Context) : UiNode(props, context) 
 
     private fun setTextSize(props: Bundle) {
         if (props.containsKey(PROP_TEXT_SIZE)) {
-            val sizeMeters = props.getDouble(PROP_TEXT_SIZE)
+            val sizeMeters = props.getDouble(PROP_TEXT_SIZE).toFloat()
             val size = Utils.metersToPx(sizeMeters, view.context).toFloat()
             (view as TextView).setTextSize(TypedValue.COMPLEX_UNIT_PX, size)
         }
